refactor: remove PluginRegistry and related types, consolidate plugin interfaces
- Deleted PluginRegistry.ts and its associated logic for managing plugins.
- Merged plugin-related types into index.ts for better organization.
- Updated index.ts to include simplified Plugin and PluginRoute interfaces.
- Removed outdated tests and utility functions related to the previous plugin structure.
- Adjusted package.json to reflect new paths for plugin types and imports.
