+++ frontend/plugins/pagoda-minimal-builder/src/generatedPlugins.ts
@@ -2,6 +2,7 @@
 // Edit plugins.config.js and run npm run build instead.
 
 import plugin0 from "pagoda-plugin-hello-world";
+import plugin1 from "pagoda-plugin-dashboard";
 
 // Export all plugins
-export const configuredPlugins = [plugin0].filter(Boolean);
+export const configuredPlugins = [plugin0, plugin1].filter(Boolean);
