+++ frontend/plugins/plugin-hello-world/package.json
@@ -4,7 +4,6 @@
   "description": "Hello World plugin for Airone - demonstrates basic plugin functionality",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
-  "type": "module",
   "files": [
     "dist",
     "src"
